Add TypeScript to Vue.js
- Add tsloader
- Add TypeScript dependency
- Move Main and Register to TypeScript
- Separate template & script for Register.vue
