+++ generators/client/templates/package.json.ejs
@@ -63,6 +63,9 @@
     "shelljs": "^0.7.6",
     "uglifyjs-webpack-plugin": "^1.1.1",
     "url-loader": "^0.5.8",
+    "ts-loader": "^3.5.0",
+    "tslib": "^1.9.3",
+    "typescript": "^3.1.3",
     "vue-jest": "^3.0.0",
     "vue-loader": "^13.3.0",
     "vue-style-loader": "^3.0.1",
