+++ generators/client/templates/tsconfig.json.ejs
@@ -1,0 +1,35 @@
+{
+    "compilerOptions": {
+        "target": "es5",
+        "esModuleInterop": true,
+        "allowSyntheticDefaultImports": true,
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "sourceMap": true,
+        "emitDecoratorMetadata": true,
+        "experimentalDecorators": true,
+        "removeComments": false,
+        "noImplicitAny": false,
+        "skipLibCheck": true,
+        "suppressImplicitAnyIndexErrors": true,
+        "outDir": "target/www/app",
+        "lib": ["es7", "dom"],
+        "typeRoots": [
+            "node_modules/@types"
+        ],
+        "baseUrl": "./",
+        "paths": {
+            "app/*": ["src/main/webapp/app/*"]
+        },
+        "importHelpers": true,
+        "allowJs": true
+    },
+    "include": [
+        "src/main/webapp/app/components",
+        "src/main/webapp/app/shims-vue.d.ts",
+        "src/main/webapp/app/main.ts"
+    ],
+    "exclude": [
+        "node_modules"
+    ]
+}
