+++ generators/client/templates/webapp/app/router/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 import Home from '../components/Home';
-import Register from '../components/account/Register';
+import Register from '../components/account/register/Register.vue';
 import ResetPassword from '../components/account/ResetPassword';
 import ChangePassword from '../components/account/ChangePassword';
 import Sessions from '../components/account/Sessions';
