+++ generators/client/templates/webpack/utils.js.ejs
@@ -1,0 +1,73 @@
+<%#
+ Copyright 2013-2018 the original author or authors from the JHipster project.
+
+ This file is part of the JHipster project, see https://jhipster.github.io/
+ for more information.
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+-%>
+const fs = require('fs');
+const path = require('path');
+
+module.exports = {
+  parseVersion,
+  root,
+  isExternalLib
+};
+
+<%_ if (buildTool === 'maven') { _%>
+const parseString = require('xml2js').parseString;
+// return the version number from `pom.xml` file
+function parseVersion() {
+  let version = null;
+  const pomXml = fs.readFileSync('pom.xml', 'utf8');
+  parseString(pomXml, (err, result) => {
+    if (result.project.version && result.project.version[0]) {
+      version = result.project.version[0];
+    } else if (result.project.parent && result.project.parent[0] && result.project.parent[0].version && result.project.parent[0].version[0]) {
+      version = result.project.parent[0].version[0];
+    }
+  });
+  if (version === null) {
+    throw new Error('pom.xml is malformed. No version is defined');
+  }
+  return version;
+}
+<%_ } else if (buildTool === 'gradle') { _%>
+// Returns the second occurrence of the version number from `build.gradle` file
+function parseVersion() {
+  const versionRegex = /^version\s*=\s*[',"]([^',"]*)[',"]/gm; // Match and group the version number
+  const buildGradle = fs.readFileSync('build.gradle', 'utf8');
+  return versionRegex.exec(buildGradle)[1];
+}
+<%_ } else { _%>
+// Returns a static version number when server is skipped
+function parseVersion() {
+  return '0.0.1-SNAPSHOT';
+}
+<%_ } _%>
+
+const _root = path.resolve(__dirname, '..');
+
+function root(args) {
+  args = Array.prototype.slice.call(arguments, 0);
+  return path.join.apply(path, [_root].concat(args));
+}
+
+function isExternalLib(module, check = /node_modules/) {
+  const req = module.userRequest;
+  if (typeof req !== 'string') {
+    return false;
+  }
+  return req.search(check) >= 0;
+}
