+++ generators/client/templates/webpack/webpack.common.js
@@ -4,7 +4,7 @@ const utils = require('./vue.utils');
 const config = require('../config');
 const vueLoaderConfig = require('./loader.conf');
 
-function resolve (dir) {
+function resolve(dir) {
     return path.join(__dirname, '..', dir);
 }
 
@@ -13,7 +13,7 @@ function resolve (dir) {
 module.exports = {
     context: path.resolve(__dirname, '../'),
     entry: {
-        app: './src/main/webapp/app/main.js'
+        app: './src/main/webapp/app/main.ts'
     },
     output: {
         path: config.build.assetsRoot,
@@ -23,7 +23,7 @@ module.exports = {
             : config.dev.assetsPublicPath
     },
     resolve: {
-        extensions: ['.js', '.vue', '.json'],
+        extensions: ['.ts', '.js', '.vue', '.json'],
         alias: {
             'vue$': 'vue/dist/vue.esm.js',
             '@': resolve('src/main/webapp/app')
@@ -39,7 +39,25 @@ module.exports = {
             {
                 test: /\.js$/,
                 loader: 'babel-loader',
-                include: [resolve('src'), resolve('test'), resolve('node_modules/webpack-dev-server/client'),resolve('bootstrap-vue')]
+                include: [resolve('src'), resolve('test'), resolve('node_modules/webpack-dev-server/client'), resolve("bootstrap-vue")]
+            },
+            {
+                test: /\.ts$/,
+                use: [
+                    {
+                        loader: 'babel-loader'
+                    },
+                    {
+                        loader: 'ts-loader',
+                        options: {
+                            appendTsSuffixTo: [
+                                '\\.vue$'
+                            ],
+                            happyPackMode: false
+                        }
+                    }
+                ],
+                include: [resolve('src'), resolve('test')]
             },
             {
                 test: /\.(png|jpe?g|gif|svg)(\?.*)?$/,
@@ -68,8 +86,8 @@ module.exports = {
         ]
     },
     node: {
-    // prevent webpack from injecting useless setImmediate polyfill because Vue
-    // source contains it (although only uses it if it's native).
+        // prevent webpack from injecting useless setImmediate polyfill because Vue
+        // source contains it (although only uses it if it's native).
         setImmediate: false,
         // prevent webpack from injecting mocks to Node native modules
         // that does not make sense for the client
